Move additional data to own table, secure models to user

+++ app/Http/Controllers/Pages/ActivityFileController.php
@@ -11,12 +11,11 @@ use Illuminate\Support\Facades\Storage;
 class ActivityFileController extends Controller
 {
 
-    public function download(Activity $activity, File $file)
-    {
-    }
-
     public function destroy(Activity $activity, File $file)
     {
+        $this->authorize('view', $file);
+        abort_if($activity->activity_file_id !== $file->id, 404, 'The file is not attached to the activity');
+
         $file->delete();
 
         return redirect()->route('activity.show', $activity);
@@ -24,6 +23,9 @@ class ActivityFileController extends Controller
 
     public function update(Request $request, Activity $activity, File $file)
     {
+        $this->authorize('update', $file);
+        abort_if($activity->activity_file_id !== $file->id, 404, 'The file is not attached to the activity');
+
         $request->validate([
             'title' => 'sometimes|nullable|string|min:1|max:254',
             'caption' => 'sometimes|nullable|string|min:1|max:60000'
